requirements.txt
asgiref==3.8.1
Django==4.2.15
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==24.2
pillow==10.4.0
pip==25.0
PyJWT==2.9.0
setuptools==75.1.0
sqlparse==0.5.3
tzdata==2025.1
pymysql
